+++ test/fixtures/my-lib.js
@@ -8,3 +8,15 @@ exports.old = function () {
 exports.old2 = function () {
   deprecate('old2')
 }
+
+exports.automsg = function () {
+  deprecate()
+}
+
+exports.automsgnamed = function automsgnamed() {
+  deprecate()
+}
+
+exports.automsganon = function () {
+  (function () { deprecate() }())
+}