+++ History.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Automatically generate message when omitted from `deprecate()`
+
 0.0.1 / 2014-06-15
 ==================